# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CarryGo backend API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      parameters:
        - in: query
          name: fromCity
          schema: { type: string }
        - in: query
          name: toCity
          schema: { type: string }
        - in: query
          name: mine
          schema: { type: boolean }
      responses:
        "200":
          description: Trips
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Trip" }
    post:
      operationId: createTrip
      tags: [trips]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTripBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trip" }
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trip" }
  /trips/{id}/cancel:
    post:
      operationId: cancelTrip
      tags: [trips]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Trip" }
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      parameters:
        - in: query
          name: direction
          schema:
            type: string
            enum: [incoming, outgoing, all]
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DeliveryRequest" }
    post:
      operationId: createRequest
      tags: [requests]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRequestBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeliveryRequest" }
  /requests/{id}/accept:
    post:
      operationId: acceptRequest
      tags: [requests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Delivery created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Delivery" }
  /requests/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [requests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeliveryRequest" }
  /requests/{id}/cancel:
    post:
      operationId: cancelRequest
      tags: [requests]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeliveryRequest" }
...
components:
  schemas:
...
    User:
      type: object
      required: [id, phone, name, avatarColor, rating, ratingsCount, joinedAt]
      properties:
        id: { type: string }
        phone: { type: string }
        name: { type: string }
        avatarColor: { type: string }
        rating: { type: number }
        ratingsCount: { type: integer }
        joinedAt: { type: string, format: date-time }
    Trip:
      type: object
      required: [id, travellerId, travellerName, travellerRating, fromCity, toCity, date, vehicle, capacityKg, status, createdAt]
      properties:
        id: { type: string }
        travellerId: { type: string }
        travellerName: { type: string }
        travellerRating: { type: number }
        travellerAvatarColor: { type: string }
        fromCity: { type: string }
        toCity: { type: string }
        date: { type: string, format: date-time }
        vehicle: { type: string, enum: [Bike, Car, Bus, Train, Flight] }
        capacityKg: { type: number }
        notes: { type: string, nullable: true }
        status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
        createdAt: { type: string, format: date-time }
    CreateTripBody:
      type: object
      required: [fromCity, toCity, date, vehicle, capacityKg]
      properties:
        fromCity: { type: string, minLength: 1 }
        toCity: { type: string, minLength: 1 }
        date: { type: string, format: date-time }
        vehicle: { type: string, enum: [Bike, Car, Bus, Train, Flight] }
        capacityKg: { type: number, minimum: 0.1 }
        notes: { type: string }
    Parcel:
      type: object
      required: [id, senderId, senderName, fromCity, toCity, date, category, description, weightKg, priceOffer, receiverName, receiverPhone, status, createdAt]
      properties:
        id: { type: string }
        senderId: { type: string }
        senderName: { type: string }
        senderAvatarColor: { type: string }
        fromCity: { type: string }
        toCity: { type: string }
        date: { type: string, format: date-time }
        category: { type: string, enum: [Documents, Electronics, Clothing, Food, Medicine, Other] }
        description: { type: string }
        weightKg: { type: number }
        priceOffer: { type: integer }
        receiverName: { type: string }
        receiverPhone: { type: string }
        imageUri: { type: string, nullable: true }
        status: { type: string, enum: [OPEN, MATCHED, DELIVERED, CANCELLED, FAILED] }
        createdAt: { type: string, format: date-time }
...
    DeliveryRequest:
      type: object
      required: [id, parcelId, tripId, senderId, travellerId, status, createdAt, updatedAt, parcel, trip, sender, traveller]
      properties:
        id: { type: string }
        parcelId: { type: string }
        tripId: { type: string }
        senderId: { type: string }
        travellerId: { type: string }
        status: { type: string, enum: [PENDING, ACCEPTED, REJECTED, CANCELLED] }
        message: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        parcel: { $ref: "#/components/schemas/Parcel" }
        trip: { $ref: "#/components/schemas/Trip" }
        sender: { $ref: "#/components/schemas/User" }
        traveller: { $ref: "#/components/schemas/User" }
    CreateRequestBody:
      type: object
      required: [parcelId, tripId]
      properties:
        parcelId: { type: string }
        tripId: { type: string }
        message: { type: string }
    Delivery:
      type: object
      required: [id, requestId, parcelId, tripId, senderId, travellerId, stage, otp, pricePaid, escrowReleased, senderRated, travellerRated, parcel, trip, sender, traveller]
      properties:
        id: { type: string }
        requestId: { type: string }
        parcelId: { type: string }
        tripId: { type: string }
        senderId: { type: string }
        travellerId: { type: string }
        stage: { type: string, enum: [AWAITING_PICKUP, IN_TRANSIT, DELIVERED, FAILED] }
        pickupConfirmedAt: { type: string, format: date-time, nullable: true }
        deliveryConfirmedAt: { type: string, format: date-time, nullable: true }
        otp: { type: string }
        pricePaid: { type: integer }
        escrowReleased: { type: boolean }
        failedReason: { type: string, nullable: true }
        senderRated: { type: boolean }
        travellerRated: { type: boolean }
        parcel: { $ref: "#/components/schemas/Parcel" }
        trip: { $ref: "#/components/schemas/Trip" }
        sender: { $ref: "#/components/schemas/User" }
        traveller: { $ref: "#/components/schemas/User" }
...