# lib/api-zod/src/generated/types/createRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CarryGo backend API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRequestBody {
  parcelId: string;
  tripId: string;
  message?: string;
}